Add files via upload
The R files help create the graphs and charts to analyze the data.

The game in 2018 broke the character classes into 4 categories - Offense, Defense, Tanks, and Healers.

SROffense.R, SRDefense.R, SRTanks.R, and SRHealers.R all help determine the pick rate differences of each character in their respective category.

overwatchdata.R is the main file that helps to run and analyze many different graphs using a data matrix and box graphs for each of the four classes to further prove the data and hypothesis. The data tables used in this file (OverwatchOffense, OverwatchDefense, OverwatchTanks, OverwatchHeals, overwatchdata, OVerwatchdataExOL, and SR) are all in the other branch labelled "Data-tables".